Changed generation of random encryption key for temporary file.
This is a theoretical enhancement that is unlikely to make any difference in practice.
The rationale is that if one of the multiplicands is zero, then the product is zero, making the key predictable.
By adding instead of multipying, the sum is zero only if both of the terms are zero.